Added tests for receiver.
